feat(tls): expand tls options
SparkPost recently deprecated TLS v1.0 on their end (seems to be a global trend
in regard to PCI compliance) and as a user of both this lib and SparkPost, I was
kinda screwed. I monkey patched things on my end to make sure my site was
working, and this commit expands upon things / future proof them a bit.

Since the TLS 1.0, 1.1 and 1.2 constants are PHP 5.6+, I added a sanity check
for that while preserving the original functionality.

I added some tests for the new scenarios but the testing suite was throwing a
ton of "Could not open SMTP port" errors, even before I had made any code
changes. Rolling the dice since I wasn't able to fully test against the test
suite.