don't use specific TLS versions anymore
Modern PHP releases do not support old and insecure TLS versions
anymore. In general it's not advised to let end users decide on the TLS
version or hardcode those versions into the library, since these things
change and you run the risk running an outdated version. For example the
code did not support TLS1.3 before this patch, because only 1.2 was
known at the time of writing the original code.

This patch ignores the passed version and will always let PHP pick the
recommended versions.